Fix DELETE error 1093: add DeleteWithSelfReferencingSubquery flag
- Add ServerVersionSupport.DeleteWithSelfReferencingSubquery (false for MySQL, true for MariaDB 11+)
- Guard 11 DELETE tests with self-referencing subqueries (Northwind 8, NonSharedModel 2, TPH 1)
- When unsupported, assert MySqlException(1093) instead of running the test

+++ test/EFCore.MySql.FunctionalTests/BulkUpdates/NonSharedModelBulkUpdatesMySqlTest.cs
@@ -4,6 +4,7 @@ using Microsoft.EntityFrameworkCore.BulkUpdates;
 using Microsoft.EntityFrameworkCore.TestUtilities;
 using MySqlConnector;
 using Pomelo.EntityFrameworkCore.MySql.FunctionalTests.TestUtilities;
+using Pomelo.EntityFrameworkCore.MySql.Tests;
 using Xunit;
 
 namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.BulkUpdates;
@@ -81,10 +82,12 @@ FROM `Owner` AS `o`
 
     public override async Task Delete_predicate_based_on_optional_navigation(bool async)
     {
-        await base.Delete_predicate_based_on_optional_navigation(async);
+        if (AppConfig.ServerVersion.Supports.DeleteWithSelfReferencingSubquery)
+        {
+            await base.Delete_predicate_based_on_optional_navigation(async);
 
-        AssertSql(
-            """
+            AssertSql(
+                """
 DELETE `p`
 FROM `Posts` AS `p`
 WHERE `p`.`Id` IN (
@@ -94,6 +97,14 @@ WHERE `p`.`Id` IN (
     WHERE `b`.`Title` LIKE 'Arthur%'
 )
 """);
+        }
+        else
+        {
+            // Not supported by MySQL and older MariaDB versions:
+            //     Error Code: 1093. You can't specify target table 'p' for update in FROM clause
+            await Assert.ThrowsAsync<MySqlException>(
+                () => base.Delete_predicate_based_on_optional_navigation(async));
+        }
     }
 
     public override async Task Update_non_owned_property_on_entity_with_owned(bool async)
@@ -158,10 +169,12 @@ SET `b0`.`Title` = CAST(`b0`.`Rating` AS char),
 
     public override async Task Delete_entity_with_auto_include(bool async)
     {
-        await base.Delete_entity_with_auto_include(async);
+        if (AppConfig.ServerVersion.Supports.DeleteWithSelfReferencingSubquery)
+        {
+            await base.Delete_entity_with_auto_include(async);
 
-        AssertSql(
-            """
+            AssertSql(
+                """
 DELETE `c`
 FROM `Context30572_Principal` AS `c`
 WHERE `c`.`Id` IN (
@@ -170,6 +183,14 @@ WHERE `c`.`Id` IN (
     LEFT JOIN `Context30572_Dependent` AS `c1` ON `c0`.`DependentId` = `c1`.`Id`
 )
 """);
+        }
+        else
+        {
+            // Not supported by MySQL and older MariaDB versions:
+            //     Error Code: 1093. You can't specify target table 'c' for update in FROM clause
+            await Assert.ThrowsAsync<MySqlException>(
+                () => base.Delete_entity_with_auto_include(async));
+        }
     }
 
     public override async Task Update_with_alias_uniquification_in_setter_subquery(bool async)
